requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
django-filter==2.4.0
PyJWT==2.1.0
flake8==5.0.4
flake8-docstrings==1.7.0
pep8-naming==0.13.3
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework-simplejwt==5.3.1
